new: [live-profile] add custom facility to propagate live environment to git and ssh
Proxy info from the system doesn't get propagated automatically, and
ssh-agent environment is not also available to git. This adds a small
wrapper that tries to bring the less overhead as possible (I had a
bash script in a first try that took 35ms per call on wen), which
translated to seconds in the 20-30 calls that magit routinely makes.

We are now roughly around 0.3ms per call most of the time.